convert longs to ints

+++ src/main/java/org/tomitribe/util/hash/XxHash32.java
@@ -13,7 +13,7 @@
  */
 package org.tomitribe.util.hash;
 
-import static java.lang.Long.rotateLeft;
+import static java.lang.Integer.rotateLeft;
 import static org.tomitribe.util.hash.JvmUtils.unsafe;
 import static org.tomitribe.util.hash.Preconditions.checkPositionIndexes;
 
@@ -22,44 +22,44 @@ import static org.tomitribe.util.hash.Preconditions.checkPositionIndexes;
  * @author Martin Traverso
  */
 public class XxHash32 {
-    private final static long PRIME32_1 = 506952113; // 2654435761L & 0x7FFFFFFF
-    private final static long PRIME32_2 = 99338871; // 2246822519L & 0x7FFFFFFF
-    private final static long PRIME32_3 = 1119006269; // 3266489917L & 0x7FFFFFFF
-    private final static long PRIME32_4 = 668265263;
-    private final static long PRIME32_5 = 374761393;
+    private final static int PRIME32_1 = 506952113; // 2654435761L & 0x7FFFFFFF
+    private final static int PRIME32_2 = 99338871; // 2246822519L & 0x7FFFFFFF
+    private final static int PRIME32_3 = 1119006269; // 3266489917L & 0x7FFFFFFF
+    private final static int PRIME32_4 = 668265263;
+    private final static int PRIME32_5 = 374761393;
 
-    private final static long DEFAULT_SEED = 0;
+    private final static int DEFAULT_SEED = 0;
 
-    public static long hash(String data) {
+    public static int hash(String data) {
         return hash(Slices.utf8Slice(data));
     }
 
-    public static long hash(Slice data) {
+    public static int hash(Slice data) {
         return hash(data, 0, data.length());
     }
 
-    public static long hash(long seed, Slice data) {
+    public static int hash(int seed, Slice data) {
         return hash(seed, data, 0, data.length());
     }
 
-    public static long hash(Slice data, int offset, int length) {
+    public static int hash(Slice data, int offset, int length) {
         return hash(DEFAULT_SEED, data, offset, length);
     }
 
-    public static long hash(long seed, Slice data, int offset, int length) {
+    public static int hash(int seed, Slice data, int offset, int length) {
         checkPositionIndexes(0, offset + length, data.length());
 
         Object base = data.getBase();
         long p = data.getAddress() + offset;
         long end = p + length;
 
-        long hash;
+        int hash;
 
         if (length >= 16) {
-            long v1 = seed + PRIME32_1 + PRIME32_2;
-            long v2 = seed + PRIME32_2;
-            long v3 = seed + 0;
-            long v4 = seed - PRIME32_1;
+            int v1 = seed + PRIME32_1 + PRIME32_2;
+            int v2 = seed + PRIME32_2;
+            int v3 = seed + 0;
+            int v4 = seed - PRIME32_1;
 
             long limit = end - 16;
             do {
@@ -108,7 +108,7 @@ public class XxHash32 {
         return hash;
     }
 
-    private static long mix(long current, long value) {
+    private static int mix(int current, int value) {
         return rotateLeft(current + value * PRIME32_2, 13) * PRIME32_1;
     }
 }
